LPS-179960 Move DynamicObjectDefinitionTable and DynamicObjectRelationshipMappingTable from service to api
