+++ modules/apps/object/object-api/src/main/java/com/liferay/object/dsl/DynamicObjectDefinitionTable.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.object.internal.petra.sql.dsl;
+package com.liferay.object.dsl;
 
 import com.liferay.object.constants.ObjectFieldConstants;
 import com.liferay.object.model.ObjectDefinition;
