+++ modules/apps/object/object-api/src/main/java/com/liferay/object/dsl/DynamicObjectRelationshipMappingTable.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.object.internal.petra.sql.dsl;
+package com.liferay.object.dsl;
 
 import com.liferay.petra.sql.dsl.Column;
 import com.liferay.petra.sql.dsl.base.BaseTable;
